# app/Http/Controllers/Api/CategoryController.php
<?php

namespace App\Http\Controllers\Api;

use App\Http\Controllers\Controller;
use App\Http\Requests\StoreCategoryRequest;
use Illuminate\Http\Request;
use App\Models\Category;
use App\Http\Resources\CategoryResource;
use Symfony\Component\HttpFoundation\Response;
use Illuminate\Support\Str;
use Symfony\Component\Console\Descriptor\Descriptor;

/**
 * @group Categories
 *
 * Managing Categories
 */
class CategoryController extends Controller
{
    /**
     * Get categories
     * 
     * Getting the list of categories
     * 
     * @queryParam page int Which page to show. Example: 12
     * @queryParam point string Which page to show. Example: "test"
     */
    public function index()
    {
        abort_if(!auth()->user->tokenCan('categories-list'), 403);
        return CategoryResource::collection(Category::all());
    }

    public function show(Category $category)
    {
        abort_if(!auth()->user->tokenCan('categories-show'), 403);
        return new CategoryResource($category);
    }

    /**
     * POST categories
     * 
     * @bodyParam name string required Name of the category. Example: "Clothing"
     */
    public function store(StoreCategoryRequest $request)
    {
        $data = $request->all();
 
        if ($request->hasFile('photo')) { 
            $file = $request->file('photo');
            $name = 'categories/' . Str::uuid() . '.' . $file->extension();
            $file->storePubliclyAs('images', $name);
            $data['photo'] = $name;
        } 
 
        $category = Category::create($data);
 
        return new CategoryResource($category);
    }

    public function update(Category $category, StoreCategoryRequest $request)
    {
        $category->update($request->all());

        return new CategoryResource($category);
    }

    public function destroy(Category $category)
    {
        $category->delete();

        return response(null, Response::HTTP_NO_CONTENT);
    }
}