# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Api-Course
  description: ''
  version: 1.0.0
servers:
  -
    url: 'https://api-course.desplega.com'
paths:
  /api/categories:
    get:
      summary: 'Get categories'
      operationId: getCategories
      description: 'Getting the list of categories'
      parameters:
        -
          in: query
          name: page
          description: 'Which page to show.'
          example: 12
          required: false
          schema:
            type: integer
            description: 'Which page to show.'
            example: 12
        -
          in: query
          name: point
          description: 'Which page to show.'
          example: '"test"'
          required: false
          schema:
            type: string
            description: 'Which page to show.'
            example: '"test"'
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Categories
      security: []
    post:
      summary: 'POST categories'
      operationId: pOSTCategories
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Categories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'Name of the category.'
                  example: '"Clothing"'
              required:
                - name
      security: []
...